+++ src/renderer/pages/Plugin.vue
@@ -179,7 +179,7 @@
       <template #footer>
         <el-button
           round
-          @click="dialogVisible = false"
+          @click="handleCancelConfig"
         >
           {{ $T('CANCEL') }}
         </el-button>
@@ -450,6 +450,11 @@ async function handleConfirmConfig () {
   }
 }
 
+function handleCancelConfig () {
+  dialogVisible.value = false
+  $configForm.value?.resetConfig()
+}
+
 function _getSearchResult (val: string) {
   // this.$http.get(`https://api.npms.io/v2/search?q=${val}`)
   axios.get(`https://registry.npmjs.com/-/v1/search?text=${val}`)
@@ -480,7 +485,7 @@ function handleSearchResult (item: INPMSearchResultObject) {
   return {
     name,
     fullName: item.package.name,
-    author: item.package.author.name,
+    author: item.package.maintainers[0]?.username || '',
     description: item.package.description,
     logo: `https://cdn.jsdelivr.net/npm/${item.package.name}/logo.png`,
     config: {},