+++ src/universal/types/rpc.d.ts
@@ -7,6 +7,8 @@ type IGetLatestVersionArgs = [isCheckBetaVersion: boolean]
 type IToolboxCheckArgs = [type: import('./enum').IToolboxItemType]
 type IOpenFileArgs = [filePath: string]
 type ICopyTextArgs = [text: string]
+type IShowDockIconArgs = [visible: boolean]
+type IGetGalleryMenuListArgs = [selectedList: IGalleryItem[]]
 
 interface IRPCServer {
   start: () => void